fix(deps): aws-sdk should not be a direct dependency
Since aws-sdk is included in the lambda runtime, often users want to exclude it from their function
packages. When aws-sdk is imported as a direct dependency then users have to manually remove it from
their node_modules directory. By switching it to a dev dep it becomes the user's choice whether they
want to fix the aws-sdk version and include it in their bundle or not.